+++ tests/conftest.py
@@ -101,8 +101,8 @@ def app(database_url, monkeypatch):
     from titiler.pgstac.main import app
 
     # Remove middlewares https://github.com/encode/starlette/issues/472
-    app.user_middleware = []
-    app.middleware_stack = app.build_middleware_stack()
+    # app.user_middleware = []
+    # app.middleware_stack = app.build_middleware_stack()
 
     with TestClient(app) as app:
         yield app